Add project files

+++ migrations/20240722165608_create_balance_histories_table.sql
@@ -1,8 +1,8 @@
 -- Add migration script here
 CREATE TABLE balance_histories (
-  id SERIAL PRIMARY KEY,
-  user_id INTEGER,
-  balance_id INTEGER,
+  id BIGSERIAL PRIMARY KEY,
+  user_id BIGINT,
+  balance_id BIGINT,
   balance DECIMAL(12, 2),
   top_up DECIMAL(12, 2) DEFAULT 0,
   created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
